src/requirements.txt
flask<3.0
gspread
google-auth
